+++ ChangeLog.md
@@ -7,6 +7,10 @@ All notable changes to `sebastianbergmann/money` will be documented in this file
 ### Added
 * Added the `Money::getConvertedAmount()` method for converting a `Money` object's amount into its base units
 * Added the `Currency::getCurrencies()` method for retrieving the registered currencies
+* Added the `IntlFormatter::fromLocale()` named constructor to create `IntlFormatter` objects for a specified locale
+
+### Changed
+* The constructor of the `IntlFormatter` class now expects a `NumberFormatter` object
 
 ### Removed
 * Removed support for PHP < 5.5.0